Entity Fixed
! Fixed : Entity location
